requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.4
django_celery_results==2.6.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
flower==2.0.1
gevent==25.5.1
greenlet==3.2.3
humanize==4.12.3
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
packaging==25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
zope.event==5.1
zope.interface==7.2
